fix sequelize

+++ models/book.js
@@ -4,7 +4,7 @@ module.exports = function(sequelize, DataTypes) {
   var Book = sequelize.define("Book", {
     name: {type:DataTypes.STRING (60) , defaultValue:"Unknown"},
     status:{type:DataTypes.INTEGER(1),defaultValue:1},
-    DailyRent : {type:DataTypes.DOUBLE(3)}
+    DailyRent : {type:DataTypes.DOUBLE()}
    }, {
     classMethods: {
       associate: function(models) {
